More work into supporting pure CSS themes. Fixed a bug where dirty polygons were unreported. UISceneNodes now are updated forcing the current UISceneNode as the one that it's being updated. Added "poly" function for background/foreground images. Added "tab-height" and "window-title" CSS properties.

+++ src/eepp/ui/uitab.cpp
@@ -116,11 +116,11 @@ UIPushButton* UITab::setText( const String& text ) {
 
 void UITab::onAutoSize() {
 	if ( mFlags & UI_AUTO_SIZE ) {
-		Uint32 w = mTextBox->getSize().getWidth() +
-				   ( NULL != mIcon ? mIcon->getSize().getWidth() + mIcon->getLayoutMargin().Left +
-										 mIcon->getLayoutMargin().Right
-								   : 0 ) +
-				   getSkinSize().getWidth();
+		Float w = mTextBox->getSize().getWidth() +
+				  ( NULL != mIcon ? mIcon->getSize().getWidth() + mIcon->getLayoutMargin().Left +
+										mIcon->getLayoutMargin().Right
+								  : 0 ) +
+				  getSkinSize().getWidth();
 
 		UITabWidget* tTabW = getTabWidget();
